CRUD and their testcases of all Entities

+++ src/main/java/com/ERP/entities/HR.java
@@ -26,7 +26,7 @@ public class HR {
             strategy = GenerationType.SEQUENCE,
             generator = "hr_sequence"
     )
-    private int hrId;
+    private long hrId;
 
     private String name;
     private String password;
